feat: remove rewire and extract functions
to follow Single Responsability Principle and simplify tests
