+++ package.json
@@ -77,7 +77,6 @@
     "node-crisp-api": "^1.12.2",
     "nodemailer": "^6.6.3",
     "pg": "^8.7.1",
-    "rewire": "^5.0.0",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5",
     "uuid": "^8.3.2",
@@ -91,7 +90,6 @@
     "@typescript-eslint/eslint-plugin": "^4.28.5",
     "@typescript-eslint/parser": "^4.28.5",
     "babel-core": "^6.26.3",
-    "babel-plugin-rewire": "^1.2.0",
     "chai": "^4.3.4",
     "chai-http": "^4.3.0",
     "ejs-lint": "^1.2.1",
