+++ services/demarchesSimplifiees/autoAccept.ts
@@ -7,7 +7,7 @@ import {
 } from '../champsAndAnnotations';
 import importDossier from './importDossier';
 import uploadDocument from './uploadDocument';
-import graphql from '../../utils/graphql';
+import graphql from './buildRequest';
 
 const FILE = path.join(
   __dirname, '..', '..', '..', 'static', 'documents', 'parcours_psychologue_sante_psy_etudiant.pdf',
