+++ services/demarchesSimplifiees/importDossier.ts
@@ -1,79 +1,7 @@
-import graphql from '../../utils/graphql';
-import uuid from '../../utils/uuid';
-import config from '../../utils/config';
-import { getChampsFieldFromId } from '../champsAndAnnotations';
+import graphql from './buildRequest';
+import parsePsychologists from './parsePsychologists';
 import { DSPsychologist, Psychologist } from '../../types/Psychologist';
 import { DSResponse } from '../../types/DemarcheSimplifiee';
-
-const parseTeleconsultation = (inputString: string): boolean => inputString === 'true';
-
-/**
- * transform string "speciality1, speciality2" to array ["speciality1", "speciality2"]
- * as a JSON to store it inside PG
- */
-const parseTraining = (inputString: string): string => {
-  if (inputString.includes(',')) {
-    return JSON.stringify(inputString.split(', '));
-  }
-  return JSON.stringify([inputString]);
-};
-
-const getUuidDossierNumber = (number: number): string => (
-  uuid.generateFromString(`${config.demarchesSimplifieesId}-${number}`)
-);
-
-const parseDossierMetadata = (dossier: DSPsychologist): Psychologist => {
-  const {
-    state,
-    archived,
-    demandeur,
-    usager,
-    number,
-    groupeInstructeur,
-    champs,
-  } = dossier;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const psy: any = { state, archived };
-
-  psy.dossierNumber = getUuidDossierNumber(number);
-
-  psy.lastName = demandeur.nom.trim();
-  psy.firstNames = demandeur.prenom.trim();
-
-  psy.personalEmail = usager.email.trim();
-
-  psy.region = groupeInstructeur.label;
-
-  const champsToMap = [
-    'departement',
-    'diploma',
-    'adeli',
-    'phone',
-    'address',
-    'email',
-    'languages',
-    'website',
-    'description',
-    'teleconsultation',
-    'training',
-  ];
-  champs.forEach((champ) => {
-    const field = getChampsFieldFromId(champ.id);
-    if (champsToMap.includes(field)) {
-      psy[field] = champ.stringValue.trim();
-    }
-  });
-  psy.teleconsultation = parseTeleconsultation(psy.teleconsultation);
-  psy.training = parseTraining(psy.training);
-
-  return psy;
-};
-
-const parsePsychologists = (psychologists: DSPsychologist[]): Psychologist[] => {
-  console.log(`Parsing ${psychologists.length} psychologists from DS API`);
-
-  return psychologists.map((psychologist) => parseDossierMetadata(psychologist));
-};
 
 /**
  * helper function called by getPsychologistList
