+++ test/autoVerifyPsychologist.spec.ts
@@ -1,74 +1,31 @@
 import sinon from 'sinon';
-import graphql from '../utils/graphql';
-import config from '../utils/config';
+import verifyPsychologist from '../services/demarchesSimplifiees/verifyPsychologist';
 import autoVerifyPsychologists from '../services/demarchesSimplifiees/autoVerify';
 
 describe('autoVerifyPsychologist', () => {
-  let executeMutationStub;
-  let unsets = [];
+  let verifyPsychologistStub;
 
   beforeEach(() => {
-    unsets = [];
-    executeMutationStub = sinon.stub();
-    unsets.push(graphql.__set__('executeMutation', executeMutationStub));
-    autoVerifyPsychologists.__Rewire__('graphql', graphql);
+    verifyPsychologistStub = sinon.stub(verifyPsychologist, 'verifyPsychologist');
   });
 
   afterEach((done) => {
-    autoVerifyPsychologists.__ResetDependency__('graphql');
-    unsets.forEach((unset) => unset());
+    verifyPsychologistStub.restore();
     done();
   });
 
-  it.skip('Should execute 4 mutations on DS', async () => {
+  // FIXME!
+  it.skip('Should call verifyPsychologist twice with correct args', async () => {
     await autoVerifyPsychologists();
 
-    sinon.assert.callCount(executeMutationStub, 4);
+    sinon.assert.callCount(verifyPsychologistStub, 2);
 
-    const DOSSIER_ID_VALID = 'RG9zc2llci00ODkwOTEw';
-    const DOSSIER_ID_INVALID = 'RG9zc2llci00Nzg5MzM2';
-    const INSTRUCTOR_ID = config.demarchesSimplifieesInstructor;
+    sinon.assert.calledWith(verifyPsychologistStub.getCall(0),
+      sinon.match(),
+      sinon.match());
 
-    sinon.assert.calledWith(executeMutationStub.getCall(0),
-      sinon.match((query) => query.includes('mutation dossierModifierAnnotationText')),
-      sinon.match({
-        input: {
-          dossierId: DOSSIER_ID_INVALID,
-          instructeurId: INSTRUCTOR_ID,
-          annotationId: 'Q2hhbXAtMTY1NzQwNQ==',
-          value: sinon.match((value) => value.startsWith("Le dossier n'a pas passé la vérification automatique le ")),
-        },
-      }));
-
-    sinon.assert.calledWith(executeMutationStub.getCall(1),
-      sinon.match((query) => query.includes('mutation dossierModifierAnnotationText')),
-      sinon.match({
-        input: {
-          dossierId: DOSSIER_ID_VALID,
-          instructeurId: INSTRUCTOR_ID,
-          annotationId: 'Q2hhbXAtMTY1NzQwNQ==',
-          value: sinon.match((value) => value.startsWith('Dossier vérifié automatiquement le ')),
-        },
-      }));
-
-    sinon.assert.calledWith(executeMutationStub.getCall(2),
-      sinon.match((query) => query.includes('mutation dossierModifierAnnotationCheckbox')),
-      sinon.match({
-        input: {
-          dossierId: DOSSIER_ID_VALID,
-          instructeurId: INSTRUCTOR_ID,
-          annotationId: 'Q2hhbXAtMTY1NzQwMw==',
-          value: true,
-        },
-      }));
-
-    sinon.assert.calledWith(executeMutationStub.getCall(3),
-      sinon.match((query) => query.includes('mutation dossierPasserEnInstruction')),
-      sinon.match({
-        input: {
-          dossierId: DOSSIER_ID_VALID,
-          instructeurId: INSTRUCTOR_ID,
-        },
-      }));
+    sinon.assert.calledWith(verifyPsychologistStub.getCall(1),
+      sinon.match(),
+      sinon.match());
   }).timeout(30000);
 });
