--- test/test-uuid.ts
@@ -1,11 +1,0 @@
-import uuid from '../utils/uuid';
-
-describe('uuid', () => {
-  describe('generateFromString', () => {
-    it('should generate a uuid from a string', async () => {
-      const result = uuid.generateFromString('my-id');
-
-      result.should.be.equal('5b8fca8a-bcfa-5a6d-a97a-72404d0b0216');
-    });
-  });
-});
