+++ utils/uuid.ts
@@ -5,7 +5,11 @@ const generateFromString = (id: string): string => uuidv5(id, config.uuidNamespa
 
 const generateRandom = (random = Math.random().toString()): string => generateFromString(random);
 
+const getUuidDossierNumber = (number: number): string => (
+  generateFromString(`${config.demarchesSimplifieesId}-${number}`)
+);
 export default {
   generateFromString,
   generateRandom,
+  getUuidDossierNumber,
 };
